added to_s and inspect methods for classes
The to_s methods just contain the name of the object (and the number
for Lines), and inspect just returns the name instead of all the
instance variables. The fact that a large recursive inspect was being
printed every time I tried to to something in irb bothered me somewhat.